% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contents.R
\name{coff_TBRFA}
\alias{coff_TBRFA}
\title{For a varaiable pair calculate TBRF coefficients based on tree coefficients list}
\usage{
coff_TBRFA(tree_coff_list, Tar_pair)
}
\arguments{
\item{tree_coff_list}{A tree coefficients list returned by tree_to_allpair_coff()}

\item{Tar_pair}{A character string of two node variable connected with "_"}
}
\value{
A data frame containing basic coefficients to calculated TBRF score
}
\description{
For a varaiable pair calculate TBRF coefficients based on tree coefficients list
}
\examples{
\dontrun{
feats <- (colnames(HODC$HODC_1)[4:9]);f <- paste(feats,collapse=' + ')
f <- paste('PhenoRespon ~',f);f <- as.formula(f);f
Train = HODC$HODC_1[which(HODC$HODC_1$Order>=2),]
rf_mtry8 <- randomForest(f , data = Train, importance = TRUE,mtry = 8)
tree_1 = as.data.frame(getTree(rf_mtry8,1))
tree_cofflist = tree_to_allpair_coff(tree_1 )
coff_TBRFA (tree_cofflist,"3_6")
}
}
