% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contents.R
\name{collapse}
\alias{collapse}
\title{Replace " " with "_" in a string}
\usage{
collapse(x)
}
\arguments{
\item{x}{A string}
}
\value{
A string replaced  " " by "_"
}
\description{
Replace " " with "_" in a string
}
\examples{
x="A B"
collapse(x)
}
