% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contents.R
\name{conds_to_route}
\alias{conds_to_route}
\title{find split variables in a single decision route}
\usage{
conds_to_route(x, tree)
}
\arguments{
\item{x}{A single decision route}

\item{tree}{A single decision tree from a randomforest model}
}
\value{
characters of split variables in a single decision route
}
\description{
find split variables in a single decision route
}
\examples{
\dontrun{
feats <- (colnames(HODC$HODC_1)[4:9]);f <- paste(feats,collapse=' + ')
f <- paste('PhenoRespon ~',f);f <- as.formula(f);f
Train = HODC$HODC_1[which(HODC$HODC_1$Order>=2),]
rf_mtry8 <- randomForest(f , data = Train, importance = TRUE,mtry = 8)
tree_1 = as.data.frame(getTree(rf_mtry8,1))
tree_1_conds = getConds(tree_1)
tree_1_sv = conds_to_route(tree_1_conds[[1]],tree_1)
}
}
