% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contents.R
\name{getConds}
\alias{getConds}
\title{extract decision routes from a tree}
\usage{
getConds(tree)
}
\arguments{
\item{tree}{A single decision tree from a randomforest model}
}
\value{
A list of decision routes from the tree
}
\description{
extract decision routes from a tree
}
\examples{
\dontrun{
feats <- (colnames(HODC$HODC_1)[4:9]);f <- paste(feats,collapse=' + ')
f <- paste('PhenoRespon ~',f);f <- as.formula(f);f
Train = HODC$HODC_1[which(HODC$HODC_1$Order>=2),]
rf_mtry8 <- randomForest(f , data = Train, importance = TRUE,mtry = 8)
tree_1 = as.data.frame(getTree(rf_mtry8,1))
tree_1_conds = getConds(tree_1)
}
}
