% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contents.R
\docType{data}
\name{HODC}
\alias{HODC}
\title{HODC}
\format{
An object of class \code{character} of length 1.
}
\source{
description
}
\usage{
HODC
}
\description{
a dataset from literature, containing 7 groups higher-order drug combination test with 22 unique drugs and 5 different cell lines for 441 tests
}
\keyword{datasets}
