% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contents.R
\name{prevCond}
\alias{prevCond}
\title{find the previous decision conditions in a tree}
\usage{
prevCond(tree, i)
}
\arguments{
\item{tree}{A single decision tree from a randomforest model}

\item{i}{The row number of a terminal node in a tree dataframe}
}
\value{
A list containing the decision condition and the row number of the previous node
}
\description{
find the previous decision conditions in a tree
}
\examples{
\dontrun{
feats <- (colnames(HODC$HODC_1)[4:9]);f <- paste(feats,collapse=' + ')
f <- paste('PhenoRespon ~',f);f <- as.formula(f);f
Train = HODC$HODC_1[which(HODC$HODC_1$Order>=2),]
rf_mtry8 <- randomForest(f , data = Train, importance = TRUE,mtry = 8)
tree_1 = as.data.frame(getTree(rf_mtry8,1))
prevCond(tree_1,which(tree_1$status== -1)[1])
}
}
