% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contents.R
\name{RF2TI}
\alias{RF2TI}
\title{Calculate TI score for a randomforest model}
\usage{
RF2TI(rf_model)
}
\arguments{
\item{rf_model}{A randomforest model}
}
\value{
A data.frame containing TI score for all variable pair
}
\description{
Calculate TI score for a randomforest model
}
\examples{

\dontrun{

feats <- (colnames(HODC$HODC_1)[4:9]);f <- paste(feats,collapse=' + ')
f <- paste('PhenoRespon ~',f);f <- as.formula(f);f
Train = HODC$HODC_1[which(HODC$HODC_1$Order>=2),]
rf_model <- randomForest(f , data = Train, importance = TRUE,mtry = 8)
RF2TI(rf_model)
}
}
