% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contents.R
\name{tree_intercoff}
\alias{tree_intercoff}
\title{Find tree decision routes and split variables inside routes, for a given variable pair, return depth of variable nodes, co-occurrence, and depth difference}
\usage{
tree_intercoff(tree, x_pair_wanted)
}
\arguments{
\item{tree}{A single decision tree from a randomforest model}

\item{x_pair_wanted}{A character string of two node variable connected with "_"}
}
\value{
A list of tree decision routes,split variables inside routes, dataframe of tree coeffients for a variable pair, names of the variable pair and tree depth
}
\description{
Find tree decision routes and split variables inside routes, for a given variable pair, return depth of variable nodes, co-occurrence, and depth difference
}
\examples{
\dontrun{
feats <- (colnames(HODC$HODC_1)[4:9]);f <- paste(feats,collapse=' + ')
f <- paste('PhenoRespon ~',f);f <- as.formula(f);f
Train = HODC$HODC_1[which(HODC$HODC_1$Order>=2),]
rf_mtry8 <- randomForest(f , data = Train, importance = TRUE,mtry = 8)
tree_1 = as.data.frame(getTree(rf_mtry8,1))
tree_intercoff(tree_1 ,"3_6")
}
}
