% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contents.R
\name{tree_to_allpair_coff}
\alias{tree_to_allpair_coff}
\title{Given a tree, for all possible variable pair, extract tree decision routes and split variables inside routes, depth of variable nodes, co-occurrences, and depth differences}
\usage{
tree_to_allpair_coff(tree)
}
\arguments{
\item{tree}{A single decision tree from a randomforest model}
}
\value{
A big list containing coefficients for all possible variable pairs, namely the tree decision routes and split variables inside routes, depth of variable nodes, co-occurrences, and depth differences
}
\description{
Given a tree, for all possible variable pair, extract tree decision routes and split variables inside routes, depth of variable nodes, co-occurrences, and depth differences
}
\examples{
\dontrun{
feats <- (colnames(HODC$HODC_1)[4:9]);f <- paste(feats,collapse=' + ')
f <- paste('PhenoRespon ~',f);f <- as.formula(f);f
Train = HODC$HODC_1[which(HODC$HODC_1$Order>=2),]
rf_mtry8 <- randomForest(f , data = Train, importance = TRUE,mtry = 8)
tree_1 = as.data.frame(getTree(rf_mtry8,1))
tree_to_allpair_coff(tree_1)
}
}
