% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contents.R
\name{tree_to_TI}
\alias{tree_to_TI}
\title{Given a tree list of randomforest model,calculated the TI score of every variable pair}
\usage{
tree_to_TI(rf_tree_list)
}
\arguments{
\item{rf_tree_list}{A list of trees from a random forest model}
}
\value{
A data.frame containing TI score of all variable pair for a randomforest model
}
\description{
Given a tree list of randomforest model,calculated the TI score of every variable pair
}
\examples{
\dontrun{
feats <- (colnames(HODC$HODC_1)[4:9]);f <- paste(feats,collapse=' + ')
f <- paste('PhenoRespon ~',f);f <- as.formula(f);f
Train = HODC$HODC_1[which(HODC$HODC_1$Order>=2),]
rf_model <- randomForest(f , data = Train, importance = TRUE,mtry = 8)
rf_ntree = rf_model$ntree
find_k_tree = function(x) {tree_to_allpair_coff(getTree(rf_model, k=x, labelVar=TRUE))}
rf_tree_list = lapply(1: rf_ntree, find_k_tree)
tree_to_TI(rf_tree_list)
}
}
